Prevent attack with invalid faction. Added because gos with type GAMEOBJECT_TYPE_SPELL_FOCUS  casting spells of linked gos with type GAMEOBJECT_TYPE_TRAP instead of spawning them.

+++ src/world/Faction.cpp
@@ -177,6 +177,9 @@ bool isAttackable(Object* objA, Object* objB, bool CheckStealth)
             return true;
     }
 
+    if (objA->m_faction == nullptr)     // no faction, no kill (added because spell_caster gos should summon a trap instead of casting the spell directly.)
+        return false;
+
     if (objA->m_faction == objB->m_faction)    // same faction can't kill each other unless in ffa pvp/duel
         return false;
